+++ pandas/tests/series/test_apply.py
@@ -627,19 +627,19 @@ class TestSeriesMap:
         expected = Series([np.nan, np.nan, "three"])
         tm.assert_series_equal(result, expected)
 
-    def test_map_abc_mapping(self):
+    def test_map_abc_mapping(self, non_mapping_dict_subclass):
         # https://github.com/pandas-dev/pandas/issues/29733
         # Check collections.abc.Mapping support as mapper for Series.map
         s = Series([1, 2, 3])
-        not_a_dictionary = tm.TestNonDictMapping({3: "three"})
+        not_a_dictionary = non_mapping_dict_subclass({3: "three"})
         result = s.map(not_a_dictionary)
         expected = Series([np.nan, np.nan, "three"])
         tm.assert_series_equal(result, expected)
 
-    def test_map_abc_mapping_with_missing(self):
+    def test_map_abc_mapping_with_missing(self, non_mapping_dict_subclass):
         # https://github.com/pandas-dev/pandas/issues/29733
         # Check collections.abc.Mapping support as mapper for Series.map
-        class NonDictMappingWithMissing(tm.TestNonDictMapping):
+        class NonDictMappingWithMissing(non_mapping_dict_subclass):
             def __missing__(self, key):
                 return "missing"
